+++ search-apis/package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
+    "test:watch": "jest --watchAll",
     "dev": "nodemon index.js"
   },
   "author": "NICE Digital (https://www.nice.org.uk/digital)",
@@ -21,6 +22,8 @@
     "jsonpath": "^1.0.2"
   },
   "devDependencies": {
-    "nodemon": "^2.0.20"
+    "jest": "^29.3.1",
+    "nodemon": "^2.0.20",
+    "supertest": "^6.3.3"
   }
 }
